# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la plataforma LogiRed de transporte de paquetes",
        "title": "LogiRed API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Devuelve lista de usuarios (solo para administradores)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener todos los usuarios (solo admin)",
                "responses": {
                    "200": {
                        "description": "lista de usuarios",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.User"
                            }
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea una cuenta (cliente o conductor) con imagen opcional",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Registrar un nuevo usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Apellido",
                        "name": "lastname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Correo electrónico",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Teléfono",
                        "name": "numberphone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha de nacimiento (YYYY-MM-DD)",
                        "name": "birthdate",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contraseña",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "1=cliente, 2=conductor",
                        "name": "user_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ciudad de trabajo (solo conductores)",
                        "name": "citywork",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Foto de perfil (jpg, png, gif)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "usuario creado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error en los datos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/profile/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener perfil completo de un usuario (con datos de conductor si aplica)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del usuario",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "perfil completo",
                        "schema": {
                            "$ref": "#/definitions/entities.UserProfile"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "usuario no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.DriverInfo": {
            "type": "object",
            "properties": {
                "citywork": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "entities.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "numberphone": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_type": {
                    "type": "integer"
                }
            }
        },
        "entities.UserProfile": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "driver_info": {
                    "$ref": "#/definitions/entities.DriverInfo"
                },
                "email": {
                    "type": "string"
                },
                "iduser": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number_phone": {
                    "type": "string"
                },
                "user_type": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Escribe \"Bearer \u003ctoken\u003e\" para autenticarte",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}